+++ src/main/java/com/squareup/kotlinpoet/CodeWriter.kt
@@ -37,7 +37,7 @@ private fun extractMemberName(part: String): String {
 internal class CodeWriter constructor(
     out: Appendable,
     private val indent: String = "  ",
-    private val memberImports: Set<String> = emptySet(),
+    private val memberImports: Map<String, Import> = emptyMap(),
     private val importedTypes: Map<String, ClassName> = emptyMap()
 ) {
   private val out = LineWrapper(out, indent, 100)
@@ -60,8 +60,8 @@ internal class CodeWriter constructor(
   var statementLine = -1
 
   init {
-    for (signature in memberImports) {
-      memberImportClassNames.add(signature.substring(0, signature.lastIndexOf('.')))
+    for ((className, _) in memberImports) {
+      memberImportClassNames.add(className.substring(0, className.lastIndexOf('.')))
     }
   }
 
@@ -86,7 +86,7 @@ internal class CodeWriter constructor(
     this.packageName = NO_PACKAGE
   }
 
-  fun pushType(type: TypeSpec)= apply {
+  fun pushType(type: TypeSpec) = apply {
     this.typeSpecStack.add(type)
   }
 
@@ -154,6 +154,9 @@ internal class CodeWriter constructor(
       if (index > 0) emit(", ")
       if (typeVariable.variance != null) {
         emit("${typeVariable.variance.keyword} ")
+      }
+      if (typeVariable.reified) {
+        emit("reified ")
       }
       emitCode("%L", typeVariable.name)
       if (typeVariable.bounds.size == 1) {
@@ -282,10 +285,14 @@ internal class CodeWriter constructor(
     if (partWithoutLeadingDot.isEmpty()) return false
     val first = partWithoutLeadingDot[0]
     if (!Character.isJavaIdentifierStart(first)) return false
-    val explicit = canonical + "." + extractMemberName(partWithoutLeadingDot)
-    val wildcard = canonical + ".*"
-    if (memberImports.contains(explicit) || memberImports.contains(wildcard)) {
-      emit(partWithoutLeadingDot)
+    val explicit = memberImports[canonical + "." + extractMemberName(partWithoutLeadingDot)]
+    if (explicit != null) {
+      if (explicit.alias != null) {
+        val memberName = extractMemberName(partWithoutLeadingDot)
+        emit(partWithoutLeadingDot.replaceFirst(memberName, explicit.alias))
+      } else {
+        emit(partWithoutLeadingDot)
+      }
       return true
     }
     return false
@@ -311,11 +318,14 @@ internal class CodeWriter constructor(
     var nameResolved = false
     var c: ClassName? = className
     while (c != null) {
-      val resolved = resolve(c.simpleName())
+      val alias = memberImports[c.canonicalName]?.alias
+      val simpleName = alias ?: c.simpleName()
+      val resolved = resolve(simpleName)
       nameResolved = resolved != null
 
       // We don't care about nullability here, as it's irrelevant for imports.
       if (resolved == c.asNonNullable()) {
+        if (alias != null) return alias
         val suffixOffset = c.simpleNames().size - 1
         return className.simpleNames().subList(suffixOffset,
             className.simpleNames().size).joinToString(".")
@@ -347,7 +357,7 @@ internal class CodeWriter constructor(
       return
     }
     val topLevelClassName = className.topLevelClassName()
-    val simpleName = topLevelClassName.simpleName()
+    val simpleName = memberImports[className.canonicalName]?.alias ?: topLevelClassName.simpleName()
     val replaced = importableTypes.put(simpleName, topLevelClassName)
     if (replaced != null) {
       importableTypes.put(simpleName, replaced) // On collision, prefer the first inserted.