+++ src/main/java/com/squareup/kotlinpoet/FileSpec.kt
@@ -47,7 +47,7 @@ class FileSpec private constructor(builder: FileSpec.Builder) {
   val packageName = builder.packageName
   val name = builder.name
   val members = builder.members.toList()
-  private val memberImports = builder.memberImports.toImmutableSet()
+  private val memberImports = builder.memberImports.associateBy(Import::qualifiedName)
   private val indent = builder.indent
 
   @Throws(IOException::class)
@@ -96,18 +96,23 @@ class FileSpec private constructor(builder: FileSpec.Builder) {
 
     codeWriter.pushPackage(packageName)
 
-    if (packageName.isNotEmpty()) {
-      codeWriter.emitCode("package %L\n", packageName)
+    val escapedPackageName = packageName.split('.')
+        .joinToString(".") { escapeIfKeyword(it) }
+
+    if (escapedPackageName.isNotEmpty()) {
+      codeWriter.emitCode("package %L\n", escapedPackageName)
       codeWriter.emit("\n")
     }
 
     val imports = codeWriter.importedTypes().values
         .map { it.canonicalName }
-        .plus(memberImports)
+        .filterNot { it in memberImports.keys }
+        .plus(memberImports.map { it.value.toString() })
 
     if (imports.isNotEmpty()) {
       for (className in imports.toSortedSet()) {
-        codeWriter.emitCode("import %L\n", className)
+        codeWriter.emitCode("import %L", className)
+        codeWriter.emit("\n")
       }
       codeWriter.emit("\n")
     }
@@ -169,7 +174,7 @@ class FileSpec private constructor(builder: FileSpec.Builder) {
       internal val name: String) {
     internal val annotations = mutableListOf<AnnotationSpec>()
     internal val comment = CodeBlock.builder()
-    internal val memberImports = sortedSetOf<String>()
+    internal val memberImports = sortedSetOf<Import>()
     internal var indent = "  "
     internal val members = mutableListOf<Any>()
 
@@ -233,17 +238,33 @@ class FileSpec private constructor(builder: FileSpec.Builder) {
         = addStaticImport(`class`.asClassName(), *names)
 
     fun addStaticImport(className: ClassName, vararg names: String) = apply {
-      check(names.isNotEmpty()) { "names array is empty" }
+      require(names.isNotEmpty()) { "names array is empty" }
+      require("*" !in names) { "Wildcard imports are not allowed" }
       for (name in names) {
-        memberImports += className.canonicalName + "." + name
+        memberImports += Import(className.canonicalName + "." + name)
       }
     }
 
     fun addStaticImport(packageName: String, vararg names: String) = apply {
-      check(names.isNotEmpty()) { "names array is empty" }
+      require(names.isNotEmpty()) { "names array is empty" }
+      require("*" !in names) { "Wildcard imports are not allowed" }
       for (name in names) {
-        memberImports += packageName + "." + name
+        memberImports += Import(packageName + "." + name)
       }
+    }
+
+    fun addAliasedImport(`class`: Class<*>, `as`: String) =
+        addAliasedImport(`class`.asClassName(), `as`)
+
+    fun addAliasedImport(`class`: KClass<*>, `as`: String) =
+        addAliasedImport(`class`.asClassName(), `as`)
+
+    fun addAliasedImport(className: ClassName, `as`: String) = apply {
+      memberImports += Import(className.canonicalName, `as`)
+    }
+
+    fun addAliasedImport(className: ClassName, memberName: String, `as`: String) = apply {
+      memberImports += Import("${className.canonicalName}.$memberName", `as`)
     }
 
     fun indent(indent: String) = apply {