+++ src/test/java/com/squareup/kotlinpoet/PropertySpecTest.kt
@@ -32,6 +32,47 @@ class PropertySpecTest {
     assertThat(prop.toString()).isEqualTo("val foo: kotlin.String by Delegates.notNull()\n")
   }
 
+  @Test fun inlineSingleAccessor() {
+    val prop = PropertySpec.builder("foo", String::class)
+        .getter(FunSpec.getterBuilder()
+            .addModifiers(KModifier.INLINE)
+            .addStatement("return %S", "foo")
+            .build())
+        .build()
+
+    assertThat(prop.toString()).isEqualTo("""
+      |val foo: kotlin.String
+      |  inline get() = "foo"
+      |""".trimMargin())
+  }
+
+  @Test fun inlineBothAccessors() {
+    val prop = PropertySpec.varBuilder("foo", String::class)
+        .getter(FunSpec.getterBuilder()
+            .addModifiers(KModifier.INLINE)
+            .addStatement("return %S", "foo")
+            .build())
+        .setter(FunSpec.setterBuilder()
+            .addModifiers(KModifier.INLINE)
+            .addParameter("value", String::class)
+            .build())
+        .build()
+
+    assertThat(prop.toString()).isEqualTo("""
+      |inline var foo: kotlin.String
+      |  get() = "foo"
+      |  set(value) {
+      |  }
+      |""".trimMargin())
+  }
+
+  @Test fun inlineForbiddenOnProperty() {
+    assertThrows<IllegalArgumentException> {
+      PropertySpec.builder("foo", String::class)
+          .addModifiers(KModifier.INLINE)
+    }
+  }
+
   @Test fun equalsAndHashCode() {
     val type = Int::class
     var a = PropertySpec.builder("foo", type).build()