+++ src/test/java/com/squareup/kotlinpoet/TypeSpecTest.kt
@@ -170,7 +170,7 @@ class TypeSpecTest {
                 .build())
             .addParameter(ParameterSpec.builder("three", String::class)
                 .addAnnotation(AnnotationSpec.builder(ClassName(tacosPackage, "Pong"))
-                    .addMember("value", "%S", "pong")
+                    .addMember("%S", "pong")
                     .build())
                 .build())
             .addParameter(ParameterSpec.builder("four", String::class)
@@ -239,11 +239,11 @@ class TypeSpecTest {
         .addFunction(FunSpec.builder("fooBar")
             .addModifiers(KModifier.PUBLIC, KModifier.ABSTRACT)
             .addAnnotation(AnnotationSpec.builder(headers)
-                .addMember("value", "%S", "Accept: application/json")
-                .addMember("value", "%S", "User-Agent: foobar")
+                .addMember("%S", "Accept: application/json")
+                .addMember("%S", "User-Agent: foobar")
                 .build())
             .addAnnotation(AnnotationSpec.builder(post)
-                .addMember("value", "%S", "/foo/bar")
+                .addMember("%S", "/foo/bar")
                 .build())
             .returns(ParameterizedTypeName.get(observable, fooBar))
             .addParameter(ParameterSpec.builder("things", ParameterizedTypeName.get(things, thing))
@@ -251,12 +251,12 @@ class TypeSpecTest {
                 .build())
             .addParameter(ParameterSpec.builder("query", ParameterizedTypeName.get(map, string, string))
                 .addAnnotation(AnnotationSpec.builder(queryMap)
-                    .addMember("encodeValues", "false")
+                    .addMember("encodeValues = %L", "false")
                     .build())
                 .build())
             .addParameter(ParameterSpec.builder("authorization", string)
                 .addAnnotation(AnnotationSpec.builder(header)
-                    .addMember("value", "%S", "Authorization")
+                    .addMember("%S", "Authorization")
                     .build())
                 .build())
             .build())
@@ -269,10 +269,10 @@ class TypeSpecTest {
         |import kotlin.collections.Map
         |
         |interface Service {
-        |  @Headers([
+        |  @Headers(
         |      "Accept: application/json",
         |      "User-Agent: foobar"
-        |  ])
+        |  )
         |  @POST("/foo/bar")
         |  fun fooBar(
         |      @Body things: Things<Thing>,
@@ -287,7 +287,7 @@ class TypeSpecTest {
     val taco = TypeSpec.classBuilder("Taco")
         .addProperty(PropertySpec.builder("thing", String::class, KModifier.PRIVATE)
             .addAnnotation(AnnotationSpec.builder(ClassName(tacosPackage, "JsonAdapter"))
-                .addMember("value", "%T::class", ClassName(tacosPackage, "Foo"))
+                .addMember("%T::class", ClassName(tacosPackage, "Foo"))
                 .build())
             .build())
         .build()
@@ -307,7 +307,7 @@ class TypeSpecTest {
     val taco = TypeSpec.classBuilder("Taco")
         .addProperty(PropertySpec.builder("thing", String::class, KModifier.PRIVATE)
             .addAnnotation(AnnotationSpec.builder(ClassName(tacosPackage, "JsonAdapter"))
-                .addMember("value", "%T::class", ClassName(tacosPackage, "Foo"))
+                .addMember("%T::class", ClassName(tacosPackage, "Foo"))
                 .useSiteTarget(AnnotationSpec.UseSiteTarget.FIELD)
                 .build())
             .build())
@@ -328,9 +328,9 @@ class TypeSpecTest {
     val someType = ClassName(tacosPackage, "SomeType")
     val taco = TypeSpec.classBuilder("Foo")
         .addAnnotation(AnnotationSpec.builder(ClassName(tacosPackage, "Something"))
-            .addMember("hi", "%T.%N", someType, "PROPERTY")
-            .addMember("hey", "%L", 12)
-            .addMember("hello", "%S", "goodbye")
+            .addMember("%T.%N", someType, "PROPERTY")
+            .addMember("%L", 12)
+            .addMember("%S", "goodbye")
             .build())
         .addModifiers(KModifier.PUBLIC)
         .build()
@@ -338,9 +338,9 @@ class TypeSpecTest {
         |package com.squareup.tacos
         |
         |@Something(
-        |    hi = SomeType.PROPERTY,
-        |    hey = 12,
-        |    hello = "goodbye"
+        |    SomeType.PROPERTY,
+        |    12,
+        |    "goodbye"
         |)
         |class Foo
         |""".trimMargin())
@@ -491,25 +491,25 @@ class TypeSpecTest {
         .addModifiers(KModifier.ABSTRACT)
         .addFunction(FunSpec.builder("throwOne")
             .addAnnotation(AnnotationSpec.builder(Throws::class)
-                .addMember("value", "%T::class", IOException::class.asClassName())
+                .addMember("%T::class", IOException::class.asClassName())
                 .build())
             .build())
         .addFunction(FunSpec.builder("throwTwo")
             .addAnnotation(AnnotationSpec.builder(Throws::class)
-                .addMember("value", "%T::class, %T::class",
+                .addMember("%T::class, %T::class",
                     IOException::class.asClassName(), ClassName(tacosPackage, "SourCreamException"))
                 .build())
             .build())
         .addFunction(FunSpec.builder("abstractThrow")
             .addModifiers(KModifier.ABSTRACT)
             .addAnnotation(AnnotationSpec.builder(Throws::class)
-                .addMember("value", "%T::class", IOException::class.asClassName())
+                .addMember("%T::class", IOException::class.asClassName())
                 .build())
             .build())
         .addFunction(FunSpec.builder("nativeThrow")
             .addModifiers(KModifier.EXTERNAL)
             .addAnnotation(AnnotationSpec.builder(Throws::class)
-                .addMember("value", "%T::class", IOException::class.asClassName())
+                .addMember("%T::class", IOException::class.asClassName())
                 .build())
             .build())
         .build()
@@ -1044,15 +1044,16 @@ class TypeSpecTest {
     val mealDeal = ClassName(tacosPackage, "MealDeal")
     val menu = TypeSpec.classBuilder("Menu")
         .addAnnotation(AnnotationSpec.builder(mealDeal)
-            .addMember("price", "%L", 500)
-            .addMember("options", "%L", AnnotationSpec.builder(option)
-                .addMember("name", "%S", "taco")
-                .addMember("meat", "%T::class", beef)
-                .build())
-            .addMember("options", "%L", AnnotationSpec.builder(option)
-                .addMember("name", "%S", "quesadilla")
-                .addMember("meat", "%T::class", chicken)
-                .build())
+            .addMember("%L = %L", "price", 500)
+            .addMember("%L = [%L, %L]", "options",
+                AnnotationSpec.builder(option)
+                    .addMember("%S", "taco")
+                    .addMember("%T::class", beef)
+                    .build(),
+                AnnotationSpec.builder(option)
+                    .addMember("%S", "quesadilla")
+                    .addMember("%T::class", chicken)
+                    .build())
             .build())
         .build()
     assertThat(toString(menu)).isEqualTo("""
@@ -1060,10 +1061,7 @@ class TypeSpecTest {
         |
         |@MealDeal(
         |    price = 500,
-        |    options = [
-        |        Option(name = "taco", meat = Beef::class),
-        |        Option(name = "quesadilla", meat = Chicken::class)
-        |    ]
+        |    options = [Option("taco", Beef::class), Option("quesadilla", Chicken::class)]
         |)
         |class Menu
         |""".trimMargin())
@@ -1376,7 +1374,7 @@ class TypeSpecTest {
 
   @Test fun annotationToString() {
     val annotation = AnnotationSpec.builder(SuppressWarnings::class)
-        .addMember("value", "%S", "unused")
+        .addMember("%S", "unused")
         .build()
     assertThat(annotation.toString()).isEqualTo("@java.lang.SuppressWarnings(\"unused\")")
   }
@@ -1601,7 +1599,7 @@ class TypeSpecTest {
     val taco = TypeSpec.classBuilder("Taco")
         .addAnnotations(Arrays.asList(
             AnnotationSpec.builder(SuppressWarnings::class)
-                .addMember("value", "%S", "unchecked")
+                .addMember("%S", "unchecked")
                 .build(),
             AnnotationSpec.builder(java.lang.Deprecated::class).build()))
         .build()
@@ -1838,7 +1836,7 @@ class TypeSpecTest {
   @Test fun nameFromUnsupportedType() {
     assertThrows<IllegalArgumentException> {
       CodeBlock.builder().add("%N", String::class)
-    }.hasMessage("expected name but was " + String::class)
+    }.hasMessageThat().isEqualTo("expected name but was " + String::class)
   }
 
   @Test fun stringFromAnything() {
@@ -1876,33 +1874,33 @@ class TypeSpecTest {
   @Test fun typeFromUnsupportedType() {
     assertThrows<IllegalArgumentException> {
       CodeBlock.builder().add("%T", "kotlin.String")
-    }.hasMessage("expected type but was kotlin.String")
+    }.hasMessageThat().isEqualTo("expected type but was kotlin.String")
   }
 
   @Test fun tooFewArguments() {
     assertThrows<IllegalArgumentException> {
       CodeBlock.builder().add("%S")
-    }.hasMessage("index 1 for '%S' not in range (received 0 arguments)")
+    }.hasMessageThat().isEqualTo("index 1 for '%S' not in range (received 0 arguments)")
   }
 
   @Test fun unusedArgumentsRelative() {
     assertThrows<IllegalArgumentException> {
       CodeBlock.builder().add("%L %L", "a", "b", "c")
-    }.hasMessage("unused arguments: expected 2, received 3")
+    }.hasMessageThat().isEqualTo("unused arguments: expected 2, received 3")
   }
 
   @Test fun unusedArgumentsIndexed() {
     assertThrows<IllegalArgumentException> {
       CodeBlock.builder().add("%1L %2L", "a", "b", "c")
-    }.hasMessage("unused argument: %3")
+    }.hasMessageThat().isEqualTo("unused argument: %3")
 
     assertThrows<IllegalArgumentException> {
       CodeBlock.builder().add("%1L %1L %1L", "a", "b", "c")
-    }.hasMessage("unused arguments: %2, %3")
+    }.hasMessageThat().isEqualTo("unused arguments: %2, %3")
 
     assertThrows<IllegalArgumentException> {
       CodeBlock.builder().add("%3L %1L %3L %1L %3L", "a", "b", "c", "d")
-    }.hasMessage("unused arguments: %2, %4")
+    }.hasMessageThat().isEqualTo("unused arguments: %2, %4")
   }
 
   @Test fun superClassOnlyValidForClasses() {
@@ -2519,7 +2517,7 @@ class TypeSpecTest {
           .addSuperclassConstructorParameter("%S", "hey")
           .addFunction(FunSpec.constructorBuilder().build())
           .build()
-    }.hasMessage(
+    }.hasMessageThat().isEqualTo(
         "types without a primary constructor cannot specify secondary constructors and superclass constructor parameters")
   }
 
@@ -2597,7 +2595,7 @@ class TypeSpecTest {
   @Test fun requiresNonKeywordName() {
     assertThrows<IllegalArgumentException> {
       TypeSpec.enumBuilder("when")
-    }.hasMessage("not a valid name: when")
+    }.hasMessageThat().isEqualTo("not a valid name: when")
   }
 
   @Test fun internalFunForbiddenInInterface() {
@@ -2607,13 +2605,13 @@ class TypeSpecTest {
       type.addFunction(FunSpec.builder("eat")
           .addModifiers(ABSTRACT, INTERNAL)
           .build())
-    }.hasMessage("modifiers [ABSTRACT, INTERNAL] must contain none of [INTERNAL, PROTECTED]")
+    }.hasMessageThat().isEqualTo("modifiers [ABSTRACT, INTERNAL] must contain none of [INTERNAL, PROTECTED]")
 
     assertThrows<IllegalArgumentException> {
       type.addFunctions(listOf(FunSpec.builder("eat")
           .addModifiers(ABSTRACT, INTERNAL)
           .build()))
-    }.hasMessage("modifiers [ABSTRACT, INTERNAL] must contain none of [INTERNAL, PROTECTED]")
+    }.hasMessageThat().isEqualTo("modifiers [ABSTRACT, INTERNAL] must contain none of [INTERNAL, PROTECTED]")
   }
 
   @Test fun privateAbstractFunForbiddenInInterface() {
@@ -2623,13 +2621,13 @@ class TypeSpecTest {
       type.addFunction(FunSpec.builder("eat")
           .addModifiers(ABSTRACT, PRIVATE)
           .build())
-    }.hasMessage("modifiers [ABSTRACT, PRIVATE] must contain none or only one of [ABSTRACT, PRIVATE]")
+    }.hasMessageThat().isEqualTo("modifiers [ABSTRACT, PRIVATE] must contain none or only one of [ABSTRACT, PRIVATE]")
 
     assertThrows<IllegalArgumentException> {
       type.addFunctions(listOf(FunSpec.builder("eat")
           .addModifiers(ABSTRACT, PRIVATE)
           .build()))
-    }.hasMessage("modifiers [ABSTRACT, PRIVATE] must contain none or only one of [ABSTRACT, PRIVATE]")
+    }.hasMessageThat().isEqualTo("modifiers [ABSTRACT, PRIVATE] must contain none or only one of [ABSTRACT, PRIVATE]")
   }
 
   @Test fun internalFunForbiddenInAnnotation() {
@@ -2639,13 +2637,13 @@ class TypeSpecTest {
       type.addFunction(FunSpec.builder("eat")
           .addModifiers(INTERNAL)
           .build())
-    }.hasMessage("ANNOTATION Taco.eat requires modifiers [PUBLIC, ABSTRACT]")
+    }.hasMessageThat().isEqualTo("ANNOTATION Taco.eat requires modifiers [PUBLIC, ABSTRACT]")
 
     assertThrows<IllegalArgumentException> {
       type.addFunctions(listOf(FunSpec.builder("eat")
           .addModifiers(INTERNAL)
           .build()))
-    }.hasMessage("ANNOTATION Taco.eat requires modifiers [PUBLIC, ABSTRACT]")
+    }.hasMessageThat().isEqualTo("ANNOTATION Taco.eat requires modifiers [PUBLIC, ABSTRACT]")
   }
 
   @Test fun basicDelegateTest() {